notebooks/requirements.txt
faker
pymongo
azure-cosmos
asyncio
aiohttp